+++ app/javascript/dashboard/routes/dashboard/settings/auditlogs/Index.vue
@@ -66,6 +66,10 @@ import { mapGetters } from 'vuex';
 import TableFooter from 'dashboard/components/widgets/TableFooter';
 import timeMixin from 'dashboard/mixins/time';
 import alertMixin from 'shared/mixins/alertMixin';
+import {
+  generateTranslationPayload,
+  translationKeys,
+} from 'dashboard/helper/auditlogHelper';
 
 export default {
   components: {
@@ -116,109 +120,22 @@ export default {
       // If agent does not exist(removed/deleted), return email from audit log
       return agentName || user_id;
     },
-    extractChangedAccountUserValues(audited_changes) {
-      let changes = [];
-      let values = [];
-      const roleMapping = {
-        0: 'agent',
-        1: 'admin',
-      };
-
-      const availabilityMapping = {
-        0: 'online',
-        1: 'offline',
-        2: 'busy',
-      };
-
-      // Iterate over roles
-      if (audited_changes.role && audited_changes.role.length) {
-        changes.push('role');
-        values.push(
-          roleMapping[audited_changes.role[audited_changes.role.length - 1]]
-        );
-      }
-
-      // Iterate over availability
-      if (audited_changes.availability && audited_changes.availability.length) {
-        changes.push('availability');
-        values.push(
-          availabilityMapping[
-            audited_changes.availability[
-              audited_changes.availability.length - 1
-            ]
-          ]
-        );
-      }
-
-      return { changes, values };
-    },
     generateLogText(auditLogItem) {
-      const agentName = this.getAgentName(auditLogItem.user_id);
+      const translationPayload = generateTranslationPayload(
+        auditLogItem,
+        this.getAgentName
+      );
       const auditableType = auditLogItem.auditable_type.toLowerCase();
       const action = auditLogItem.action.toLowerCase();
-      const auditId = auditLogItem.auditable_id;
       let logActionKey = `${auditableType}:${action}`;
 
-      const translationPayload = {
-        agentName,
-        id: auditId,
-      };
-
-      if (auditableType === 'accountuser') {
-        if (action === 'update') {
-          if (auditLogItem.user_id === auditLogItem.auditable.user_id) {
-            logActionKey += ':self';
-          } else {
-            logActionKey += ':other';
-            translationPayload.user = this.getAgentName(
-              auditLogItem.auditable.user_id
-            );
-          }
-          translationPayload.invitee = this.getAgentName(
-            auditLogItem.auditable.user_id
-          );
-
-          const accountUserChanges = this.extractChangedAccountUserValues(
-            auditLogItem.audited_changes
-          );
-          translationPayload.attributes = accountUserChanges.changes;
-          translationPayload.values = accountUserChanges.values;
-        }
-
-        if (action === 'create') {
-          translationPayload.invitee = this.getAgentName(
-            auditLogItem.audited_changes.user_id
-          );
-          if (auditLogItem.audited_changes.role === 0) {
-            translationPayload.role = 'admin';
-          } else {
-            translationPayload.role = 'agent';
-          }
+      if (auditableType === 'accountuser' && action === 'update') {
+        if (auditLogItem.user_id === auditLogItem.auditable.user_id) {
+          logActionKey += ':self';
+        } else {
+          logActionKey += ':other';
         }
       }
-
-      const translationKeys = {
-        'automationrule:create': `AUDIT_LOGS.AUTOMATION_RULE.ADD`,
-        'automationrule:update': `AUDIT_LOGS.AUTOMATION_RULE.EDIT`,
-        'automationrule:destroy': `AUDIT_LOGS.AUTOMATION_RULE.DELETE`,
-        'webhook:create': `AUDIT_LOGS.WEBHOOK.ADD`,
-        'webhook:update': `AUDIT_LOGS.WEBHOOK.EDIT`,
-        'webhook:destroy': `AUDIT_LOGS.WEBHOOK.DELETE`,
-        'inbox:create': `AUDIT_LOGS.INBOX.ADD`,
-        'inbox:update': `AUDIT_LOGS.INBOX.EDIT`,
-        'inbox:destroy': `AUDIT_LOGS.INBOX.DELETE`,
-        'user:sign_in': `AUDIT_LOGS.USER_ACTION.SIGN_IN`,
-        'user:sign_out': `AUDIT_LOGS.USER_ACTION.SIGN_OUT`,
-        'team:create': `AUDIT_LOGS.TEAM.ADD`,
-        'team:update': `AUDIT_LOGS.TEAM.EDIT`,
-        'team:destroy': `AUDIT_LOGS.TEAM.DELETE`,
-        'macro:create': `AUDIT_LOGS.MACRO.ADD`,
-        'macro:update': `AUDIT_LOGS.MACRO.EDIT`,
-        'macro:destroy': `AUDIT_LOGS.MACRO.DELETE`,
-        'accountuser:create': `AUDIT_LOGS.ACCOUNT_USER.ADD`,
-        'accountuser:update:self': `AUDIT_LOGS.ACCOUNT_USER.EDIT.SELF`,
-        'accountuser:update:other': `AUDIT_LOGS.ACCOUNT_USER.EDIT.OTHER`,
-      };
 
       return this.$t(translationKeys[logActionKey] || '', translationPayload);
     },
